+++ pyg_lib/csrc/ops/matmul.cpp
@@ -4,6 +4,7 @@
 #include <torch/library.h>
 #include <torch/script.h>
 
+#include "pyg_lib/csrc/utils/check.h"
 #include "pyg_lib/csrc/utils/convert.h"
 
 namespace pyg {
@@ -13,10 +14,29 @@ namespace {
 
 std::vector<at::Tensor> _grouped_matmul(const std::vector<at::Tensor>& input,
                                         const std::vector<at::Tensor>& other) {
-  // TODO (matthias) Add TensorArg definitions.
-  // TODO (matthias) Add dispatcher support.
+  TORCH_CHECK(
+      input.size() == other.size(),
+      "Number of 'input' tensors must match number of 'other' tensors.");
+  const auto n_tensors = input.size();
+  std::vector<at::TensorArg> input_args;
+  std::vector<at::TensorArg> other_args;
+  pyg::utils::fill_tensor_args(input_args, input, "input", 0);
+  pyg::utils::fill_tensor_args(other_args, other, "other", 1);
+  at::CheckedFrom c{"grouped_matmul"};
+
+  at::checkAllDefined(c, input_args);
+  at::checkAllDefined(c, other_args);
+  at::checkAllSameType(c, input_args);
+  at::checkAllSameType(c, other_args);
+  at::checkSameType(c, input_args[0], other_args[0]);
+  for (size_t i = 0; i < n_tensors; ++i) {
+    at::checkDim(c, input_args[i], 2);
+    at::checkDim(c, other_args[i], 2);
+    at::checkSize(c, other_args[i], 0, input_args[i]->size(-1));
+  }
+
   static auto op = c10::Dispatcher::singleton()
-                       .findSchemaOrThrow("pyg::cuda_grouped_matmul", "")
+                       .findSchemaOrThrow("pyg::grouped_matmul", "")
                        .typed<decltype(_grouped_matmul)>();
   return op.call(input, other);
 }
@@ -24,7 +44,19 @@ std::vector<at::Tensor> _grouped_matmul(const std::vector<at::Tensor>& input,
 at::Tensor _segment_matmul(const at::Tensor& input,
                            const at::Tensor& ptr,
                            const at::Tensor& other) {
-  // TODO (matthias) Add TensorArg definitions.
+  at::TensorArg input_arg{input, "input", 0};
+  at::TensorArg ptr_arg{ptr, "ptr", 1};
+  at::TensorArg other_arg{other, "other", 2};
+  at::CheckedFrom c{"segment_matmul"};
+
+  at::checkAllDefined(c, {input_arg, ptr_arg, other_arg});
+  at::checkSameType(c, input_arg, other_arg);
+  at::checkDim(c, input_arg, 2);
+  at::checkDim(c, ptr_arg, 1);
+  at::checkDim(c, other_arg, 3);
+  at::checkSize(c, other_arg, 1, input_arg->size(-1));
+  at::checkNumel(c, ptr_arg, other_arg->size(0) + 1);
+
   static auto op = c10::Dispatcher::singleton()
                        .findSchemaOrThrow("pyg::segment_matmul", "")
                        .typed<decltype(_segment_matmul)>();