+++ pyg_lib/ops/__init__.py
@@ -7,11 +7,8 @@ from torch import Tensor
 class SegmentMatmul(torch.autograd.Function):
     @staticmethod
     def forward(ctx, inputs, ptr, other):
-        assert inputs.is_cuda
-        assert ptr.is_cuda
-        assert other.is_cuda
         ctx.save_for_backward(inputs, ptr, other)
-        return torch.ops.pyg.cuda_segment_matmul(inputs, ptr, other)
+        return torch.ops.pyg.segment_matmul(inputs, ptr, other)
 
     @staticmethod
     def backward(ctx, out_grad):
@@ -19,7 +16,7 @@ class SegmentMatmul(torch.autograd.Function):
 
         input_grad = None
         if inputs.requires_grad:
-            input_grad = torch.ops.pyg.cuda_segment_matmul(
+            input_grad = torch.ops.pyg.segment_matmul(
                 out_grad, ptr, torch.transpose(other, -2, -1))
 
         other_grad = None, None
@@ -40,11 +37,8 @@ class SegmentMatmul(torch.autograd.Function):
 class GroupedMatmul(torch.autograd.Function):
     @staticmethod
     def forward(ctx, inputs: List[Tensor], others: List[Tensor]):
-        for x, other in zip(inputs, others):
-            assert x.is_cuda
-            assert other.is_cuda
         ctx.save_for_backward(inputs, others)
-        outs = torch.ops.pyg.cuda_grouped_matmul(inputs, others)
+        outs = torch.ops.pyg.grouped_matmul(inputs, others)
 
         # NOTE Autograd doesnt set out[i].requires_grad = True automatically
         for i in range(len(outs)):
@@ -60,13 +54,13 @@ class GroupedMatmul(torch.autograd.Function):
         if all([x.requires_grad for x in inputs]):
             for i in range(len(others)):
                 others[i] = others[i].t()
-            inputs_grad = torch.ops.pyg.cuda_grouped_matmul(outs_grad, others)
+            inputs_grad = torch.ops.pyg.grouped_matmul(outs_grad, others)
 
         others_grad = None
         if all([other.requires_grad for other in others]):
             for i in range(len(inputs)):
                 inputs[i] = inputs[i].t()
-            others_grad = torch.ops.pyg.cuda_grouped_matmul(inputs, outs_grad)
+            others_grad = torch.ops.pyg.grouped_matmul(inputs, outs_grad)
 
         return inputs_grad, others_grad